api: Move most of agent-specific modules into API/Agent
