+++ hercules-ci-agent/hercules-ci-agent/Hercules/Agent.hs
@@ -12,14 +12,15 @@ import           Control.Concurrent.Async.Lifted
                                                 ( replicateConcurrently_ )
 import           Control.Concurrent.MVar.Lifted ( withMVar )
 
-import           Hercules.API                   ( tasksReady
+import           Hercules.API                   ( noContent
+                                                )
+import           Hercules.API.Agent.Tasks                   ( tasksReady
                                                 , tasksSetStatus
-                                                , noContent
                                                 )
 import           Hercules.API.Task              ( Task )
 import qualified Hercules.API.Task             as Task
 import qualified Hercules.API.TaskStatus       as TaskStatus
-import qualified Hercules.API.EvaluateTask     as EvaluateTask
+import qualified Hercules.API.Agent.Evaluate.EvaluateTask     as EvaluateTask
 import qualified Hercules.API.Agent.Build.BuildTask
                                                as BuildTask
 import           Hercules.Agent.Client          ( tasksClient )
@@ -76,7 +77,7 @@ main = Init.setupLogging $ \logEnv -> do
               x -> pure x
             )
           $ Env.runHerculesClient
-          $ Hercules.API.tasksReady tasksClient
+          $ tasksReady tasksClient
 
         forM_ taskMaybe $ performTask
 
@@ -88,7 +89,7 @@ performTask task =
         (\e -> do
           logLocM ErrorS $ "Exception in task: " <> show (e :: SomeException)
           retry (cap 60 exponential) $ noContent $ runHerculesClient $
-            Hercules.API.tasksSetStatus tasksClient
+            tasksSetStatus tasksClient
                                         (Task.id task)
                                         (TaskStatus.Exceptional $ show e)
         )
@@ -109,6 +110,6 @@ performTask task =
         logLocM InfoS "Completed task successfully"
 
         noContent $ runHerculesClient $
-          Hercules.API.tasksSetStatus tasksClient
+          tasksSetStatus tasksClient
                                       (Task.id task)
                                       (TaskStatus.Successful ())
