+++ hercules-ci-agent/hercules-ci-agent/Hercules/Agent/NixPath.hs
@@ -7,7 +7,7 @@ where
 
 import           Protolude
 import qualified Data.Text                     as T
-import qualified Hercules.API.EvaluateTask     as EvaluateTask
+import qualified Hercules.API.Agent.Evaluate.EvaluateTask     as EvaluateTask
 
 renderNixPath :: [EvaluateTask.NixPathElement (EvaluateTask.SubPathOf FilePath)]
               -> Text
