+++ hercules-ci-api/src/Hercules/API/Agent/AgentTask.hs
@@ -1,0 +1,9 @@
+{-# LANGUAGE DeriveAnyClass #-}
+module AgentTask where
+
+import           Hercules.API.Prelude
+import qualified Hercules.API.Agent.Evaluate.EvaluateTask     as EvaluateTask
+
+data AgentTask = Evaluate EvaluateTask.EvaluateTask
+               | Build Text
+  deriving (Generic, Show, Eq, ToJSON, FromJSON, ToSchema)
