+++ tests/agent-test/hercules-ci-agent-test.cabal
@@ -23,6 +23,7 @@ source-repository head
 executable hercules-ci-agent-test
   main-is: Main.hs
   other-modules:
+      AgentTask
       DummyApi
       EvaluationSpec
       MockTasksApi
